+++ packages/rockets-server-auth/src/domains/invitation/modules/rockets-auth-invitation-acceptance.module.ts
@@ -34,4 +34,4 @@ export class RocketsAuthInvitationAcceptanceModule extends RocketsAuthInvitation
 }
 
 // Re-export for convenience
-export { type InvitationAcceptedEventHandler } from './rockets-auth-invitation-acceptance.module-definition';
+export { type InvitationUserOnboardingServiceClass } from './rockets-auth-invitation-acceptance.module-definition';